IoT/Mirror/video_requirement.txt
absl-py==1.4.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.1
contourpy==1.1.0
cycler==0.11.0
decorator==4.4.2
dlib==19.24.2
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==23.5.26
fonttools==4.42.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-resources==6.0.1
ipython==8.12.2
jedi==0.19.0
joblib==1.3.1
kiwisolver==1.4.4
lazy-loader==0.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mediapipe==0.10.3
moviepy==1.0.3
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.8.0.74
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.13
pycparser==2.21
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
sounddevice==0.4.6
stack-data==0.6.2
threadpoolctl==3.2.0
tifffile==2023.7.10
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
websockets==11.0.3
zipp==3.16.2
